fix: fix user address related

+++ src/main/java/com/store/controller/store/vo/AStoreUserAddressVO.java
@@ -12,7 +12,7 @@ public class AStoreUserAddressVO {
 
     private String userPhone;
 
-    private Byte defaultFlag;
+    private Integer defaultFlag;
 
     private String provinceName;
